Split config CLI extension for file budget
