+++ CLI/CMUXCLI+Config.swift
@@ -1,0 +1,446 @@
+import Foundation
+
+extension CMUXCLI {
+    func runConfigCommand(
+        commandArgs: [String],
+        socketPath: String,
+        explicitPassword: String?,
+        jsonOutput: Bool
+    ) throws {
+        let parsedArgs = docsSettingsArguments(commandArgs)
+        let wantsJSON = jsonOutput || parsedArgs.head.contains("--json")
+        let args = parsedArgs.arguments
+        let subcommand = args.first?.lowercased() ?? "help"
+
+        if hasHelpRequest(beforeSeparator: parsedArgs.head) {
+            print(configUsage())
+            return
+        }
+
+        switch subcommand {
+        case "help":
+            print(configUsage())
+        case "path", "paths":
+            guard args.count == 1 else {
+                throw CLIError(message: "Usage: cmux config path")
+            }
+            printSettingsPaths(jsonOutput: wantsJSON)
+        case "docs", "documentation":
+            guard args.count == 1 else {
+                throw CLIError(message: "Usage: cmux config docs")
+            }
+            try runDocsCommand(commandArgs: ["settings"], jsonOutput: wantsJSON)
+        case "doctor", "check", "validate":
+            let doctorArgs = Array(args.dropFirst())
+            let report = try runConfigDoctor(arguments: doctorArgs, jsonOutput: wantsJSON)
+            if report.errorCount > 0 {
+                throw CLIError(message: "cmux config doctor found \(report.errorCount) error(s)")
+            }
+        case "reload":
+            guard args.count == 1 else {
+                throw CLIError(message: "Usage: cmux config reload")
+            }
+            let client = try connectClient(
+                socketPath: socketPath,
+                explicitPassword: explicitPassword,
+                launchIfNeeded: false
+            )
+            defer { client.close() }
+            let response = try client.send(command: "reload_config")
+            if response.hasPrefix("ERROR:") {
+                throw CLIError(message: response)
+            }
+            print(response)
+        default:
+            throw CLIError(message: "Unknown config subcommand '\(subcommand)'. Run 'cmux config --help'.")
+        }
+    }
+
+    func configCommandDoesNotNeedSocket(_ commandArgs: [String]) -> Bool {
+        let parsedArgs = docsSettingsArguments(commandArgs)
+        let subcommand = parsedArgs.arguments.first?.lowercased() ?? "help"
+        return hasHelpRequest(beforeSeparator: parsedArgs.head) ||
+            ["help", "path", "paths", "docs", "documentation", "doctor", "check", "validate"].contains(subcommand)
+    }
+
+    func configUsage() -> String {
+        return """
+        Usage: cmux config <doctor|path|docs|reload>
+
+        Inspect cmux.json, print configuration references, or reload the running app.
+
+        Subcommands:
+          doctor [--path <path>]   Validate JSONC syntax for cmux config files.
+          path                     Print cmux.json paths, docs URL, and schema URL.
+          docs                     Print the same output as `cmux docs settings`.
+          reload                   Alias for `cmux reload-config`.
+
+        Config files:
+          \(Self.primarySettingsDisplayPath)
+          legacy config: \(Self.legacySettingsDisplayPath)
+          legacy app support: \(Self.fallbackSettingsDisplayPath)
+
+        Examples:
+          cmux config doctor
+          cmux config doctor --path .cmux/cmux.json
+          cmux config reload
+        """
+    }
+
+    func printSettingsPaths(jsonOutput: Bool) {
+        let payload: [String: Any] = [
+            "primary": Self.primarySettingsDisplayPath,
+            "legacy": Self.legacySettingsDisplayPath,
+            "fallback": Self.fallbackSettingsDisplayPath,
+            "docs_url": Self.settingsDocsURL,
+            "schema_url": Self.settingsSchemaURL,
+            "reload_command": "cmux reload-config",
+            "backup": "Back up any existing cmux.json file to a timestamped .bak copy before editing so the user can revert.",
+        ]
+
+        if jsonOutput {
+            print(jsonString(payload))
+            return
+        }
+
+        print("Config files:")
+        print("  primary:  \(Self.primarySettingsDisplayPath)")
+        print("  legacy config: \(Self.legacySettingsDisplayPath)")
+        print("  legacy app support: \(Self.fallbackSettingsDisplayPath)")
+        print()
+        print("Docs:")
+        print("  \(Self.settingsDocsURL)")
+        print()
+        print("Schema:")
+        print("  \(Self.settingsSchemaURL)")
+        print()
+        print("Before editing cmux.json:")
+        print("  Back up any existing cmux.json file to a timestamped .bak copy so the user can revert.")
+        print()
+        print("After editing cmux.json:")
+        print("  cmux reload-config")
+    }
+
+    private struct ConfigDoctorOptions {
+        let paths: [String]
+    }
+
+    private struct ConfigDoctorTarget {
+        let label: String
+        let displayPath: String
+        let path: String
+        let missingIsError: Bool
+    }
+
+    private struct ConfigDoctorFinding {
+        let label: String
+        let displayPath: String
+        let path: String
+        let status: String
+        let message: String?
+        let keys: [String]
+        let byteCount: Int?
+
+        var isError: Bool { status == "error" }
+
+        var payload: [String: Any] {
+            var result: [String: Any] = [
+                "label": label,
+                "display_path": displayPath,
+                "path": path,
+                "status": status,
+                "ok": !isError,
+                "keys": keys,
+            ]
+            if let message {
+                result["message"] = message
+            }
+            if let byteCount {
+                result["bytes"] = byteCount
+            }
+            return result
+        }
+    }
+
+    private struct ConfigDoctorReport {
+        let findings: [ConfigDoctorFinding]
+
+        var errorCount: Int {
+            findings.filter(\.isError).count
+        }
+
+        var payload: [String: Any] {
+            [
+                "ok": errorCount == 0,
+                "error_count": errorCount,
+                "findings": findings.map(\.payload),
+                "reload_command": "cmux reload-config",
+                "docs_url": CMUXCLI.settingsDocsURL,
+                "schema_url": CMUXCLI.settingsSchemaURL,
+            ]
+        }
+    }
+
+    private func runConfigDoctor(arguments: [String], jsonOutput: Bool) throws -> ConfigDoctorReport {
+        let options = try parseConfigDoctorOptions(arguments)
+        let targets = options.paths.isEmpty
+            ? defaultConfigDoctorTargets()
+            : options.paths.enumerated().map { index, rawPath in
+                let path = Self.absoluteConfigPath(rawPath)
+                return ConfigDoctorTarget(
+                    label: "custom \(index + 1)",
+                    displayPath: Self.tildePath(path),
+                    path: path,
+                    missingIsError: true
+                )
+            }
+        let findings = targets.map(configDoctorFinding(for:))
+        let report = ConfigDoctorReport(findings: findings)
+
+        if jsonOutput {
+            print(jsonString(report.payload))
+        } else {
+            printConfigDoctorReport(report)
+        }
+        return report
+    }
+
+    private func parseConfigDoctorOptions(_ arguments: [String]) throws -> ConfigDoctorOptions {
+        var paths: [String] = []
+        var index = 0
+        while index < arguments.count {
+            let argument = arguments[index]
+            if argument == "--path" {
+                let nextIndex = index + 1
+                guard nextIndex < arguments.count else {
+                    throw CLIError(message: "cmux config doctor --path requires a path")
+                }
+                paths.append(arguments[nextIndex])
+                index += 2
+                continue
+            }
+            if argument.hasPrefix("--path=") {
+                let rawPath = String(argument.dropFirst("--path=".count))
+                guard !rawPath.isEmpty else {
+                    throw CLIError(message: "cmux config doctor --path requires a path")
+                }
+                paths.append(rawPath)
+                index += 1
+                continue
+            }
+            if argument.hasPrefix("-") {
+                throw CLIError(message: "Unknown config doctor option '\(argument)'")
+            }
+            throw CLIError(message: "Unknown config doctor argument '\(argument)'. Use --path <path>.")
+        }
+        return ConfigDoctorOptions(paths: paths)
+    }
+
+    private func defaultConfigDoctorTargets() -> [ConfigDoctorTarget] {
+        let primary = Self.absoluteConfigPath(Self.primarySettingsDisplayPath)
+        var targets = [
+            ConfigDoctorTarget(
+                label: "primary",
+                displayPath: Self.primarySettingsDisplayPath,
+                path: primary,
+                missingIsError: false
+            )
+        ]
+
+        if let projectPath = findProjectConfigPath(), projectPath != primary {
+            targets.append(
+                ConfigDoctorTarget(
+                    label: "project",
+                    displayPath: Self.tildePath(projectPath),
+                    path: projectPath,
+                    missingIsError: false
+                )
+            )
+        }
+
+        let optionalPaths = [
+            ("legacy config", Self.legacySettingsDisplayPath),
+            ("legacy app support", Self.fallbackSettingsDisplayPath),
+        ]
+        for (label, displayPath) in optionalPaths {
+            let path = Self.absoluteConfigPath(displayPath)
+            guard path != primary,
+                  FileManager.default.fileExists(atPath: path),
+                  !targets.contains(where: { $0.path == path }) else {
+                continue
+            }
+            targets.append(
+                ConfigDoctorTarget(
+                    label: label,
+                    displayPath: displayPath,
+                    path: path,
+                    missingIsError: false
+                )
+            )
+        }
+        return targets
+    }
+
+    private func findProjectConfigPath() -> String? {
+        let fileManager = FileManager.default
+        let rawHomePath = ProcessInfo.processInfo.environment["HOME"] ?? fileManager.homeDirectoryForCurrentUser.path
+        let homePath = URL(fileURLWithPath: rawHomePath).standardizedFileURL.path
+        var current = URL(fileURLWithPath: fileManager.currentDirectoryPath).standardizedFileURL.path
+        while true {
+            if current == homePath {
+                return nil
+            }
+            let candidates = [
+                ((current as NSString).appendingPathComponent(".cmux") as NSString)
+                    .appendingPathComponent("cmux.json"),
+                (current as NSString).appendingPathComponent("cmux.json"),
+            ]
+            for candidate in candidates {
+                var isDirectory = ObjCBool(false)
+                if fileManager.fileExists(atPath: candidate, isDirectory: &isDirectory),
+                   !isDirectory.boolValue {
+                    return URL(fileURLWithPath: candidate).standardizedFileURL.path
+                }
+            }
+            let parent = (current as NSString).deletingLastPathComponent
+            if parent == current {
+                return nil
+            }
+            current = parent
+        }
+    }
+
+    private func configDoctorFinding(for target: ConfigDoctorTarget) -> ConfigDoctorFinding {
+        let fileManager = FileManager.default
+        guard fileManager.fileExists(atPath: target.path) else {
+            let message = target.missingIsError
+                ? "file not found"
+                : "not found; cmux will use defaults until this file exists"
+            return ConfigDoctorFinding(
+                label: target.label,
+                displayPath: target.displayPath,
+                path: target.path,
+                status: target.missingIsError ? "error" : "missing",
+                message: message,
+                keys: [],
+                byteCount: nil
+            )
+        }
+
+        do {
+            let data = try Data(contentsOf: URL(fileURLWithPath: target.path))
+            guard !data.isEmpty else {
+                return ConfigDoctorFinding(
+                    label: target.label,
+                    displayPath: target.displayPath,
+                    path: target.path,
+                    status: "error",
+                    message: "file is empty",
+                    keys: [],
+                    byteCount: 0
+                )
+            }
+            let sanitized = try JSONCParser.preprocess(data: data)
+            let object = try JSONSerialization.jsonObject(with: sanitized)
+            guard let dictionary = object as? [String: Any] else {
+                return ConfigDoctorFinding(
+                    label: target.label,
+                    displayPath: target.displayPath,
+                    path: target.path,
+                    status: "error",
+                    message: "top-level value must be a JSON object",
+                    keys: [],
+                    byteCount: data.count
+                )
+            }
+            return ConfigDoctorFinding(
+                label: target.label,
+                displayPath: target.displayPath,
+                path: target.path,
+                status: "ok",
+                message: "JSONC syntax is valid",
+                keys: dictionary.keys.sorted(),
+                byteCount: data.count
+            )
+        } catch {
+            return ConfigDoctorFinding(
+                label: target.label,
+                displayPath: target.displayPath,
+                path: target.path,
+                status: "error",
+                message: Self.configDoctorErrorMessage(error),
+                keys: [],
+                byteCount: nil
+            )
+        }
+    }
+
+    private func printConfigDoctorReport(_ report: ConfigDoctorReport) {
+        print("cmux config doctor")
+        for finding in report.findings {
+            print("\(finding.status.uppercased()) \(finding.label): \(finding.displayPath)")
+            print("  path: \(finding.path)")
+            if let byteCount = finding.byteCount {
+                print("  bytes: \(byteCount)")
+            }
+            if !finding.keys.isEmpty {
+                print("  keys: \(finding.keys.joined(separator: ", "))")
+            }
+            if let message = finding.message {
+                print("  \(message)")
+            }
+        }
+        print()
+        print("Docs: \(Self.settingsDocsURL)")
+        print("Schema: \(Self.settingsSchemaURL)")
+        print("Reload: cmux reload-config")
+    }
+
+    private static func absoluteConfigPath(_ rawPath: String) -> String {
+        let homePath = ProcessInfo.processInfo.environment["HOME"] ?? NSHomeDirectory()
+        let expanded: String
+        if rawPath == "~" {
+            expanded = homePath
+        } else if rawPath.hasPrefix("~/") {
+            expanded = (homePath as NSString).appendingPathComponent(String(rawPath.dropFirst(2)))
+        } else {
+            expanded = rawPath
+        }
+
+        let absolute = (expanded as NSString).isAbsolutePath
+            ? expanded
+            : (FileManager.default.currentDirectoryPath as NSString).appendingPathComponent(expanded)
+        return URL(fileURLWithPath: absolute).standardizedFileURL.path
+    }
+
+    private static func tildePath(_ path: String) -> String {
+        let homePath = URL(fileURLWithPath: ProcessInfo.processInfo.environment["HOME"] ?? NSHomeDirectory())
+            .standardizedFileURL
+            .path
+        let normalized = URL(fileURLWithPath: path).standardizedFileURL.path
+        if normalized == homePath {
+            return "~"
+        }
+        let prefix = homePath.hasSuffix("/") ? homePath : homePath + "/"
+        if normalized.hasPrefix(prefix) {
+            return "~/" + String(normalized.dropFirst(prefix.count))
+        }
+        return normalized
+    }
+
+    private static func configDoctorErrorMessage(_ error: Error) -> String {
+        let nsError = error as NSError
+        if let debug = nsError.userInfo[NSDebugDescriptionErrorKey] as? String {
+            let trimmed = debug.trimmingCharacters(in: .whitespacesAndNewlines)
+            if !trimmed.isEmpty {
+                return trimmed
+            }
+        }
+        let localized = nsError.localizedDescription.trimmingCharacters(in: .whitespacesAndNewlines)
+        if !localized.isEmpty {
+            return localized
+        }
+        let described = String(describing: error).trimmingCharacters(in: .whitespacesAndNewlines)
+        return described.isEmpty ? "unknown config parse error" : described
+    }
+}
